service-python/dashboardUser/requirements.txt
Flask==0.12.1
Flask-SQLAlchemy==2.2
Mako==1.0.6
packaging==16.8
gunicorn==19.9.0
psycopg2==2.7.7
pyparsing==2.2.0
python-editor==1.0.3
six==1.10.0
pymongo==3.11.0
python-dotenv==0.10.3
numpy==1.16.6
Flask-OpenTracing==1.1.0
jaeger-client==4.4.0
sentry-sdk[flask]==0.19.5
pyjwt==1.0.0
redis==3.5.3
pytz
